# osutil/mount: optimize flagOptSearch some more