--- a/osutil/mount/mount_linux.go
+++ b/osutil/mount/mount_linux.go
@@ -20,54 +20,71 @@
 package mount
 
 import (
+	"math/bits"
 	"syscall"
 )
 
-type syscallNumPair struct {
-	str string
-	num int
+type flagInfo struct {
+	mask uint32
+	name string
+}
+
+func knownMask(knownFlags []flagInfo) uint32 {
+	var mask uint32
+	for _, fi := range knownFlags {
+		mask |= fi.mask
+	}
+	return mask
 }
 
 // UMOUNT_NOFOLLOW is not defined in go's syscall package
 const UMOUNT_NOFOLLOW = 8
 
-var mountSyscalls = []syscallNumPair{
-	{"MS_REMOUNT", syscall.MS_REMOUNT},
-	{"MS_BIND", syscall.MS_BIND},
-	{"MS_REC", syscall.MS_REC},
-	{"MS_RDONLY", syscall.MS_RDONLY},
-	{"MS_SHARED", syscall.MS_SHARED},
-	{"MS_SLAVE", syscall.MS_SLAVE},
-	{"MS_PRIVATE", syscall.MS_PRIVATE},
-	{"MS_UNBINDABLE", syscall.MS_UNBINDABLE},
+var mountFlags = []flagInfo{
+	{name: "MS_REMOUNT", mask: syscall.MS_REMOUNT},
+	{name: "MS_BIND", mask: syscall.MS_BIND},
+	{name: "MS_REC", mask: syscall.MS_REC},
+	{name: "MS_RDONLY", mask: syscall.MS_RDONLY},
+	{name: "MS_SHARED", mask: syscall.MS_SHARED},
+	{name: "MS_SLAVE", mask: syscall.MS_SLAVE},
+	{name: "MS_PRIVATE", mask: syscall.MS_PRIVATE},
+	{name: "MS_UNBINDABLE", mask: syscall.MS_UNBINDABLE},
 }
 
-var unmountSyscalls = []syscallNumPair{
-	{"UMOUNT_NOFOLLOW", UMOUNT_NOFOLLOW},
-	{"MNT_FORCE", syscall.MNT_FORCE},
-	{"MNT_DETACH", syscall.MNT_DETACH},
-	{"MNT_EXPIRE", syscall.MNT_EXPIRE},
+var mountFlagsMask = knownMask(mountFlags)
+
+var unmountFlags = []flagInfo{
+	{name: "UMOUNT_NOFOLLOW", mask: UMOUNT_NOFOLLOW},
+	{name: "MNT_FORCE", mask: syscall.MNT_FORCE},
+	{name: "MNT_DETACH", mask: syscall.MNT_DETACH},
+	{name: "MNT_EXPIRE", mask: syscall.MNT_EXPIRE},
 }
 
-func flagOptSearch(flags int, lookupTable []syscallNumPair) (opts []string, unknown int) {
-	var f, i int
-	var sys syscallNumPair
-	for i, sys = range lookupTable {
-		f = sys.num
-		if flags&f == f {
-			flags ^= f
-			opts = append(opts, lookupTable[i].str)
+var unmountFlagsMask = knownMask(unmountFlags)
+
+func flagOptSearch(flags int, knownFlags []flagInfo, knownMask uint32) (opts []string, unknown int) {
+	var f uint32 = uint32(flags)
+	for _, fi := range knownFlags {
+		if f&knownMask == 0 {
+			break
+		}
+		if f&fi.mask != 0 {
+			if opts == nil {
+				opts = make([]string, 0, bits.OnesCount32(f))
+			}
+			f ^= fi.mask
+			opts = append(opts, fi.name)
 		}
 	}
-	return opts, flags
+	return opts, int(f)
 }
 
 // MountFlagsToOpts returns the symbolic representation of mount flags.
 func MountFlagsToOpts(flags int) (opts []string, unknown int) {
-	return flagOptSearch(flags, mountSyscalls)
+	return flagOptSearch(flags, mountFlags, mountFlagsMask)
 }
 
 // UnmountFlagsToOpts returns the symbolic representation of unmount flags.
 func UnmountFlagsToOpts(flags int) (opts []string, unknown int) {
-	return flagOptSearch(flags, unmountSyscalls)
+	return flagOptSearch(flags, unmountFlags, unmountFlagsMask)
 }